requirements.txt
fastapi==0.109.0
uvicorn==0.27.0
biopython==1.83
numpy==1.26.3
pandas==2.2.0
scipy==1.12.0
requests==2.31.0
python-multipart==0.0.6
pydantic==2.5.3
pydantic-settings==2.1.0
redis==5.0.1
openai==1.12.0
anthropic==0.18.0
torch==2.2.0
torch-geometric==2.5.0
scikit-learn==1.4.0
plotly==5.20.0
nglview==3.1.2